whatsapp_bot/requirements.txt
flask
requests
google-genai
python-dotenv
gunicorn